payment_service/backend/requirements.txt
astroid==1.6.4
captain-pycard==0.9.11
Cerberus==1.2
click==6.7
Flask==1.0.2
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pika==0.12.0
psycopg2==2.7.4
pylint==1.9.1
schedule==0.5.0
six==1.11.0
Werkzeug==0.14.1
wrapt==1.10.11
yapf==0.22.0
